Add a common Node interface to the graphs pkg.
Since both the `Graph` struct on the `graphs` pkg and the `Search()` function
on the `algorithms` pkg will depend on a node representation, the `Node`
interface was refactored so that it can be used by both packages.

The interface is exported by the `graphs` pkg, whilst the implementation
is private.
Also did some DRYing on the `graphs` tests.