Simplify the `graphs` package usage
There are no planes for having a different implementation of neither
`Graph` nor `Node`, neither is there the need to use this package
outside of this particular code base.

For that reason, the interfaces were removed, as well as the redundant
getters and setters and the `Graph` and `Node` types will be used
directly. The `NodeID` type was also removed and the `uint64` type will
be used instead.